src/main/webapp/app/entities/doctor/update/doctor-update.component.spec.ts
jest.mock('@angular/router');

import { ComponentFixture, TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { FormBuilder } from '@angular/forms';
import { ActivatedRoute } from '@angular/router';
import { of, Subject } from 'rxjs';

import { DoctorService } from '../service/doctor.service';
import { IDoctor, Doctor } from '../doctor.model';

import { IUser } from 'app/entities/user/user.model';
import { UserService } from 'app/entities/user/user.service';

import { DoctorUpdateComponent } from './doctor-update.component';

describe('Component Tests', () => {
  describe('Doctor Management Update Component', () => {
    let comp: DoctorUpdateComponent;
    let fixture: ComponentFixture<DoctorUpdateComponent>;
    let activatedRoute: ActivatedRoute;
    let doctorService: DoctorService;
    let userService: UserService;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
        declarations: [DoctorUpdateComponent],
        providers: [FormBuilder, ActivatedRoute],
      })
        .overrideTemplate(DoctorUpdateComponent, '')
        .compileComponents();

      fixture = TestBed.createComponent(DoctorUpdateComponent);
      activatedRoute = TestBed.inject(ActivatedRoute);
      doctorService = TestBed.inject(DoctorService);
      userService = TestBed.inject(UserService);

      comp = fixture.componentInstance;
    });

    describe('ngOnInit', () => {
      it('Should call User query and add missing value', () => {
        const doctor: IDoctor = { id: 456 };
        const internalUser: IUser = { id: 56880 };
        doctor.internalUser = internalUser;

        const userCollection: IUser[] = [{ id: 25564 }];
        jest.spyOn(userService, 'query').mockReturnValue(of(new HttpResponse({ body: userCollection })));
        const additionalUsers = [internalUser];
        const expectedCollection: IUser[] = [...additionalUsers, ...userCollection];
        jest.spyOn(userService, 'addUserToCollectionIfMissing').mockReturnValue(expectedCollection);

        activatedRoute.data = of({ doctor });
        comp.ngOnInit();

        expect(userService.query).toHaveBeenCalled();
        expect(userService.addUserToCollectionIfMissing).toHaveBeenCalledWith(userCollection, ...additionalUsers);
        expect(comp.usersSharedCollection).toEqual(expectedCollection);
      });

      it('Should update editForm', () => {
        const doctor: IDoctor = { id: 456 };
        const internalUser: IUser = { id: 15668 };
        doctor.internalUser = internalUser;

        activatedRoute.data = of({ doctor });
        comp.ngOnInit();

        expect(comp.editForm.value).toEqual(expect.objectContaining(doctor));
        expect(comp.usersSharedCollection).toContain(internalUser);
      });
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', () => {
        // GIVEN
        const saveSubject = new Subject<HttpResponse<Doctor>>();
        const doctor = { id: 123 };
        jest.spyOn(doctorService, 'update').mockReturnValue(saveSubject);
        jest.spyOn(comp, 'previousState');
        activatedRoute.data = of({ doctor });
        comp.ngOnInit();

        // WHEN
        comp.save();
        expect(comp.isSaving).toEqual(true);
        saveSubject.next(new HttpResponse({ body: doctor }));
        saveSubject.complete();

        // THEN
        expect(comp.previousState).toHaveBeenCalled();
        expect(doctorService.update).toHaveBeenCalledWith(doctor);
        expect(comp.isSaving).toEqual(false);
      });

      it('Should call create service on save for new entity', () => {
        // GIVEN
        const saveSubject = new Subject<HttpResponse<Doctor>>();
        const doctor = new Doctor();
        jest.spyOn(doctorService, 'create').mockReturnValue(saveSubject);
        jest.spyOn(comp, 'previousState');
        activatedRoute.data = of({ doctor });
        comp.ngOnInit();

        // WHEN
        comp.save();
        expect(comp.isSaving).toEqual(true);
        saveSubject.next(new HttpResponse({ body: doctor }));
        saveSubject.complete();

        // THEN
        expect(doctorService.create).toHaveBeenCalledWith(doctor);
        expect(comp.isSaving).toEqual(false);
        expect(comp.previousState).toHaveBeenCalled();
      });

      it('Should set isSaving to false on error', () => {
        // GIVEN
        const saveSubject = new Subject<HttpResponse<Doctor>>();
        const doctor = { id: 123 };
        jest.spyOn(doctorService, 'update').mockReturnValue(saveSubject);
        jest.spyOn(comp, 'previousState');
        activatedRoute.data = of({ doctor });
        comp.ngOnInit();

        // WHEN
        comp.save();
        expect(comp.isSaving).toEqual(true);
        saveSubject.error('This is an error!');

        // THEN
        expect(doctorService.update).toHaveBeenCalledWith(doctor);
        expect(comp.isSaving).toEqual(false);
        expect(comp.previousState).not.toHaveBeenCalled();
      });
    });

    describe('Tracking relationships identifiers', () => {
      describe('trackUserById', () => {
        it('Should return tracked User primary key', () => {
          const entity = { id: 123 };
          const trackResult = comp.trackUserById(0, entity);
          expect(trackResult).toEqual(entity.id);
        });
      });
    });
  });
});
